Implement AllRoutes component
AllRoutes component is rendered at every route in the application, and
conditionally renders child components based on the current pathname.
For example, if the application is currently at '/join', the 'Join'
component will be rendered. If the application is at '/profile', the
profile component will be rendered (if the user is logged in)